lib/protocol, lib/model: Implement high precision time stamps (fixes #3305)
This renames the existing FileInfo.modified field to modified_s while
retaining the same meaning (UNIX epoch seconds), and adds a new field
int32 modified_ns that holds the nanoseconds within the second.

When scanning, modified_ns is populated from Stat() data when it's
provided by the OS/filesystem. When syncing, we set the timestamp
with the full precision.

This also adds a new advanced config option mtimeWindowS which is the
time window within which we consider to timestamps equal. It defaults to
one second which is the same behavior as before - a file with timestamp
1234567890.000000000 is considered the same as the same file with a
timestamp 123456789.123456789.

The window is special as the purpose is to handle filesystems truncating
timestamps - 1234567890.000000000 is not considered equal to
1234567889.999999999 with the default settings even though the time
difference is only a nanosecond, as they become different when both
are truncated to the 1s window.

This config can be set to 2s to account for FAT32 shenanigans.

At least on Solaris+ZFS there's an interesting issue where we can read
and understand nanosecond precision timestamps but when set by Chtimes
the precision gets truncated to microseconds. Not sure if this is a Go
bug or a Solaris bug or what, but the integration test accounts for it.